requirements.txt
djangoajax
django
django-wysiwyg-redactor
